Add KFX exporter with Entity Assembler architecture
Implement KfxExporter using a schema-driven approach that cleanly
separates element semantics (Schema/Micro) from entity packaging
(Assembler/Macro).

Key design decisions:
- Three-fragment rule: each chapter produces Content ($145),
  Storyline ($259), and Section ($260) entities
- Two-pass architecture: Pass 1 surveys IR and builds position map
  without generating Ion; Pass 2 synthesizes Ion with resolved links
- Side-effect text extraction: tokens_to_ion pushes text to accumulator
  while returning clean Ion structure with symbolic content references
- All entity links use interned symbols, not hardcoded IDs

New modules:
- src/export/kfx.rs: KfxExporter with build_chapter_entities assembler
- src/kfx/context.rs: ExportContext for symbol table, ID generation
- src/kfx/fragment.rs: KfxFragment with Ion/Raw data variants
- src/kfx/serialization.rs: Binary container serialization (CONT header)

Extended:
- src/kfx/ion.rs: Add IonWriter for binary Ion output
- src/kfx/storyline.rs: Add ir_to_tokens and tokens_to_ion for export
- src/kfx/schema.rs: Add export_attributes() for bidirectional transforms

Also fixes symbol table max_id to use KFX_MAX_SYMBOL_ID (851) instead
of count, resolving symbol resolution mismatches.

+++ src/import/kfx.rs
@@ -25,7 +25,7 @@ use crate::kfx::symbols::KfxSymbol;
 /// Shorthand for getting a KfxSymbol as u32 for field lookups.
 macro_rules! sym {
     ($variant:ident) => {
-        KfxSymbol::$variant as u32
+        KfxSymbol::$variant as u64
     };
 }
 
@@ -388,7 +388,7 @@ impl KfxImporter {
     }
 
     /// Recursively parse nav entries into a tree structure.
-    fn parse_nav_entries(&self, container: &[(u32, IonValue)]) -> Vec<TocEntry> {
+    fn parse_nav_entries(&self, container: &[(u64, IonValue)]) -> Vec<TocEntry> {
         let mut entries = Vec::new();
 
         if let Some(entry_list) = get_field(container, sym!(Entries)).and_then(|v| v.as_list()) {
@@ -579,7 +579,7 @@ impl KfxImporter {
     }
 
     /// Extract section names from a reading order struct.
-    fn extract_sections(&self, order_fields: &[(u32, IonValue)]) -> Option<Vec<String>> {
+    fn extract_sections(&self, order_fields: &[(u64, IonValue)]) -> Option<Vec<String>> {
         let sections = get_field(order_fields, sym!(Sections))?.as_list()?;
         let mut section_names = Vec::new();
         for section in sections {
